refactor: rename and move base classes to infrastructure
