+++ cognee/tasks/entity_completion/entity_extractors/dummy_entity_extractor.py
@@ -1,7 +1,7 @@
 from typing import List
 
 from cognee.modules.engine.models import Entity, EntityType
-from cognee.tasks.entity_completion.entity_extractors.base_entity_extractor import (
+from cognee.infrastructure.entities.BaseEntityExtractor import (
     BaseEntityExtractor,
 )
